Activating save changes button
