+++ apps/mobile-mzima-client/src/app/profile/information/information.page.html
@@ -4,7 +4,7 @@
     [photo]="userPhoto"
     (uploadStarted)="onUploadStarted()"
     (uploadCompleted)="onUploadCompleted()"
-    (photoChanged)="handlePhotoChange($event)"
+    (photoSelected)="handlePhotoSelected($event)"
   ></app-profile-photo>
   <!-- (photoChanged)="handlePhotoChange($event)" -->
 
@@ -56,7 +56,7 @@
           expand="inline"
           type="button"
           *ngIf="form"
-          [disabled]="form.disabled || form.invalid || form.pristine "
+          [disabled]="form.disabled || (form.invalid && !isPhotoChanged) || (!form.dirty && !isPhotoChanged)"
           (buttonClick)="updateUserInformation()"
         >
           {{"app.save_changes" | translate}}
